+++ integrations/github/src/sync.ts
@@ -70,6 +70,20 @@ export async function triggerImport(
 
     const config = getSpaceConfigOrThrow(spaceInstallation);
 
+    if (config.migrated && typeof config.migrated === 'string') {
+        logger.info(
+            `skipping import for space ${spaceInstallation.space} as it has been migrated to native installation (${config.migrated})`,
+        );
+        return;
+    }
+
+    if (config.stale) {
+        logger.info(
+            `skipping export for space ${spaceInstallation.space} as it has been marked as stale`,
+        );
+        return;
+    }
+
     if (!config.key) {
         logger.info(`No configuration found for space ${spaceId}, skipping import`);
         return;
@@ -131,6 +145,20 @@ export async function triggerExport(
             : spaceInstallation.space.id;
 
     const config = getSpaceConfigOrThrow(spaceInstallation);
+
+    if (config.migrated && typeof config.migrated === 'string') {
+        logger.info(
+            `skipping import for space ${spaceInstallation.space} as it has been migrated to native installation (${config.migrated})`,
+        );
+        return;
+    }
+
+    if (config.stale) {
+        logger.info(
+            `skipping export for space ${spaceInstallation.space} as it has been marked as stale`,
+        );
+        return;
+    }
 
     if (!config.key) {
         logger.info(`No configuration found for space ${spaceId}, skipping export`);